+++ AdventOfCode/Year2021/IDay.cs
@@ -3,7 +3,7 @@
     internal interface IDay
     {
         public string ExampleInput { get; }
-        public int SolvePart1(string input);
-        public int SolvePart2(string input);
+        public long SolvePart1(string input);
+        public long SolvePart2(string input);
     }
 }
